DreamyFinal/backend/requirements.txt
django
djangorestframework
django-jazzmin
django-cors-headers
djangorestframework-simplejwt
